Fix mistakes after merge
