+++ timesheet/src/main/java/com/gev/timesheet/controller/MyController.java
@@ -1,9 +1,8 @@
 
 package com.gev.timesheet.controller;
 
-
-
-import com.gev.timesheet.dao.EmployeeDAO;
+import com.gev.timesheet.dao.PtoDAO;
+import com.gev.timesheet.dao.PtoPolicyDAO;
 import com.gev.timesheet.dao.TimesheetDAO;
 import com.gev.timesheet.domain.Timesheet;
 import io.swagger.annotations.Api;
@@ -13,10 +12,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
 @RestController
 @Api(value="My Rest Endpoint")
@@ -46,4 +42,6 @@ public class MyController {
     @ApiOperation(value = "List all timesheets", response = Iterable.class)
     public List<Timesheet> getAllTimesheet() {
         return timesheetDAO.findAll();
-    }}
+    }
+
+}
