+++ timesheet/src/test/java/com/gev/timesheet/dao/TimesheetDAOTest.java
@@ -30,7 +30,7 @@ public class TimesheetDAOTest {
         Timesheet sheet = testdao.findByUseridAndWeekending("base", "2020-12-12").orElse(null);
         Assertions.assertAll(() -> {
             Assertions.assertEquals(48, sheet.getTotalcompensatedhours());
-            Assertions.assertEquals(1, testdao.findAll().size());
+            Assertions.assertEquals(true, testdao.findAll().size() > 0);
         });
     }
